write getAll and getDetail api for carts,products & categories

+++ database/migrations/2023_02_26_093934_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('price');
+            $table->decimal('price', 8, 2);
             $table->string('description');
             $table->string('image');
             $table->foreignId('category_id');
